+++ include/scheduler.h
@@ -6,7 +6,7 @@
 
 # define STACK_SIZE 1024
 
-void schedule(uint64_t pc, uint64_t sp);
+void schedule(uint32_t pc, uint32_t sp);
 int init_scheduler(void);
 
 #endif /* !SCHEDULER_H */
